#1632
- Fichero adjunto

+++ ripea-service-intf/src/main/java/es/caib/ripea/service/intf/base/model/FileReference.java
@@ -16,6 +16,6 @@ public class FileReference {
 	protected final String name;
 	protected final byte[] content;
 	protected final String contentType;
-	protected final int contentLength;
+	protected final Long contentLength;
 
 }
